// src/app/modules/task/enums/task-stacks.enum.ts
export enum TechStackEnum {
  JAVASCRIPT = 'JavaScript',
  TYPESCRIPT = 'TypeScript',
  HTML = 'HTML',
  CSS = 'CSS',
  ANGULAR = 'Angular',
  REACT = 'React',
  VUE = 'Vue',
  NODEJS = 'Node.js',
  NESTJS = 'NestJS',
  PYTHON = 'Python',
  DJANGO = 'Django',
  JAVA = 'Java',
  SPRING = 'Spring',
  CSHARP = 'C#',
  DOTNET = '.NET',
  PHP = 'PHP',
  LARAVEL = 'Laravel',
  RUBY = 'Ruby',
  GO = 'Go',
  SWIFT = 'Swift',
  KOTLIN = 'Kotlin',
}